build: removes  visibility from grpc_collection_subscription_lib
