+++ test/common/upstream/BUILD
@@ -97,8 +97,6 @@ envoy_cc_test(
         "//source/extensions/clusters/original_dst:original_dst_cluster_lib",
         "//source/extensions/clusters/static:static_cluster_lib",
         "//source/extensions/clusters/strict_dns:strict_dns_cluster_lib",
-        "//source/extensions/config_subscription/grpc:grpc_collection_subscription_lib",
-        "//source/extensions/config_subscription/grpc:grpc_subscription_lib",
         "//source/extensions/health_checkers/http:health_checker_lib",
         "//source/extensions/health_checkers/tcp:health_checker_lib",
         "//source/extensions/load_balancing_policies/cluster_provided:config",
